Revert "相关节点存入redis"
This reverts commit f369150b0090868cf2e058fc75231971e5f76587.

+++ unityplatform-micromainpage/mainpage-domain/src/main/java/com/zhikuntech/intellimonitor/mainpage/domain/golden/GoldenUtil.java
@@ -173,14 +173,14 @@ public class GoldenUtil {
         return value;
     }
 
-    public int getInteger(int id, String dateTime) throws Exception {
+    public double getInteger(int id, String dateTime) throws Exception {
         check();
         ServerImpl server = pool.getServerImpl();
         Historian historian = new HistorianImpl(server);
         Date date = DateUtil.stringToDate(dateTime);
         double value = historian.getIntSingleValue(id, date, RtdbHisMode.RTDB_PREVIOUS).getValue();
         server.close();
-        return (int)value;
+        return value;
     }
 
     /**
